requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.27.2
PyYAML==6.0.2
botbuilder-core==4.14.4
botbuilder-schema==4.14.4
